renaming files using the KEBAB_CASE convention
