+++ web/src/components/layout/footer.tsx
@@ -1,4 +1,4 @@
-import { Container } from './Container';
+import { Container } from './container';
 import { Link } from 'react-router-dom';
 
 export function Footer() {
