+++ web/src/components/layout/header.tsx
@@ -1,4 +1,4 @@
-import { Container } from './Container';
+import { Container } from './container';
 import { Link } from 'react-router-dom';
 
 export function Header() {
