+++ web/src/components/layout/page-wrapper.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Header } from './Header';
-import { Footer } from './Footer';
+import { Header } from './header';
+import { Footer } from './footer';
 import { cn } from '@/lib/utils.ts';
 
 interface PageWrapperProps {
