+++ web/src/components/ui/index.ts
@@ -1,10 +1,10 @@
-export { Button } from './Button';
-export { Input } from './Input';
-export { Textarea } from './Textarea';
-export { Select } from './Select';
-export { Checkbox, Radio } from './Checkbox';
-export { Card, CardHeader, CardTitle, CardBody, CardFooter } from './Card';
-export { Badge } from './Badge';
-export { Alert } from './Alert';
-export { Spinner } from './Spinner';
-export { Modal } from './Modal';
+export { Button } from './button';
+export { Input } from './input';
+export { Textarea } from './textarea';
+export { Select } from './select';
+export { Checkbox, Radio } from './checkbox';
+export { Card, CardHeader, CardTitle, CardBody, CardFooter } from './card';
+export { Badge } from './badge';
+export { Alert } from './alert';
+export { Spinner } from './spinner';
+export { Modal } from './modal';
