+++ web/src/services/event-categories-service.ts
@@ -1,4 +1,4 @@
-import apiClient from '@/lib/apiClient';
+import apiClient from '@/lib/api-client';
 import type { Category, LinkCategoryRequest } from '@/types/category';
 
 export const EventCategoriesService = {
