+++ web/src/services/events-service.ts
@@ -1,4 +1,4 @@
-import apiClient from '@/lib/apiClient';
+import apiClient from '@/lib/api-client';
 import type {
   Event,
   EventDetail,
