+++ web/src/services/superadmin-service.ts
@@ -1,4 +1,4 @@
-import apiClient from '@/lib/apiClient';
+import apiClient from '@/lib/api-client';
 import {
   type CreateEventPayload,
   type UpdateUserPayload,
